[FLINK-35931][table-planner] Use a separate test class for the regexp family of functions
